Player and NPC Interactions fully implemented (v1.0)
The player now has 4 NPC's to interact with. He can go up to each of them and press 'F' to add them to his Match List
By pressing G, the player enters a dialogue in the console that prompts them to make matches with the NPC's they have
This will all be done using a UI by next week, so this is temporary for the time being, however the skeleton code for everything is
there for changing.