Improved Dialogue
Added a new dialogue to cycle through once trust points accumulate.(Trust points endlessly increment currently. I'll fix this later)
It will quiz you, and then if your choice matches the answer, it'll raise trust points.
The new dialogue will simply give out all of the information necessary to the player.

Fixed:
Images work as intended for the moment.
Instead of an ugly pink box, we now have a nice little prompt for each npc.

Errors:
When exiting the interaction zone without talking to NPC it'll throw an error, because it is trying to close a dialogue box that doesn't exist I think. Kevin can you take a look at this? It looks like it'll be an easy fix, but I can't figure out how.